[VAS-737] Refactored code
