+++ src/pages/commisionalBundles/detail/CommissionBundleDetailPage.tsx
@@ -1,22 +1,15 @@
 import {
   Grid,
   Typography,
-  Paper,
   Stack,
   Breadcrumbs,
   Button,
-  Drawer,
-  Divider,
-  IconButton,
-  Alert,
 } from '@mui/material';
-import CloseIcon from '@mui/icons-material/Close';
-import { ButtonNaked } from '@pagopa/mui-italia';
 import { Box } from '@mui/system';
 import { TitleBox, useErrorDispatcher, useLoading } from '@pagopa/selfcare-common-frontend';
 import { Link, generatePath, useParams } from 'react-router-dom';
 import { useTranslation, TFunction } from 'react-i18next';
-import { Dispatch, SetStateAction, useEffect, useState } from 'react';
+import {  useEffect, useState } from 'react';
 import ROUTES from '../../../routes';
 import { getCommissionBundleDetails } from '../../../services/__mocks__/bundleService';
 import { useAppSelector } from '../../../redux/hooks';
@@ -27,204 +20,11 @@ import { Bundle } from '../../../api/generated/portal/Bundle';
 import SideMenu from '../../../components/SideMenu/SideMenu';
 import { TypeEnum } from '../../../api/generated/portal/BundleRequest';
 import {
-  formatBooleanValueToYesOrNo,
-  formatCurrencyEur,
-  formatDateToDDMMYYYY,
   formatDateToDDMMYYYYhhmm,
 } from '../../../utils/common-utils';
+import CommissionBundleDetailConfiguration from './components/CommissionBundleDetailConfiguration';
+import CommissionBundleDetailTaxonomies from './components/CommissionBundleDetailTaxonomies';
 
-const PaddedDrawer = ({
-  openDrawer,
-  setOpenDrawer,
-  children,
-}: {
-  openDrawer: boolean;
-  setOpenDrawer: Dispatch<SetStateAction<boolean>>;
-  children: React.ReactNode;
-}) => (
-  <Drawer open={openDrawer} onClose={() => setOpenDrawer(false)} anchor="right">
-    <Box p={3} pt={1} sx={{ minWidth: '400px' }}>
-      <Box display="flex" justifyContent="flex-end" mb={1}>
-        <IconButton onClick={() => setOpenDrawer(false)}>
-          <CloseIcon />
-        </IconButton>
-      </Box>
-      {children}
-    </Box>
-  </Drawer>
-);
-
-const bundleConfigurationFields = {
-  col1: [
-    ['paymentType', 'commissionBundlesPage.addEditCommissionBundle.form.paymentType'],
-    ['touchpoint', 'commissionBundlesPage.addEditCommissionBundle.form.touchpoint'],
-    ['paymentAmount', 'commissionBundlesPage.addEditCommissionBundle.form.commission'],
-  ],
-  col2: [
-    ['minPaymentAmount', 'commissionBundlesPage.addEditCommissionBundle.form.minImport'],
-    ['maxPaymentAmount', 'commissionBundlesPage.addEditCommissionBundle.form.maxImport'],
-    ['idBrokerPsp', 'commissionBundlesPage.addEditCommissionBundle.form.brokerCode'],
-  ],
-  col3: [
-    ['idChannel', 'commissionBundlesPage.addEditCommissionBundle.form.channelCode'],
-    ['digitalStamp', 'commissionBundlesPage.addEditCommissionBundle.form.paymentWithDigitalStamp'],
-    [
-      'digitalStampRestriction',
-      'commissionBundlesPage.addEditCommissionBundle.form.paymentOnlyDigitalStamp',
-    ],
-  ],
-  col4: [
-    ['validityDateFrom', 'commissionBundlesPage.list.headerFields.startDate'],
-    ['validityDateTo', 'commissionBundlesPage.list.headerFields.endDate'],
-    ['lastUpdatedDate', 'commissionBundlesPage.commissionBundleDetail.lastChange'],
-  ],
-  // TODO updatedBy/"Modificato da" (API doesn't retrieve this info)
-};
-const formatConfigValues = (value: any, t: TFunction<'translation'>) => {
-  if (typeof value === 'string') {
-    return value;
-  }
-  if (typeof value === 'boolean') {
-    return formatBooleanValueToYesOrNo(value, t);
-  }
-  if (typeof value === 'number') {
-    return formatCurrencyEur(value);
-  }
-  if (typeof value === 'object') {
-    return formatDateToDDMMYYYY(value);
-  }
-  return '';
-};
-const BundleConfigurationDetails = ({ bundleDetail }: { bundleDetail: Bundle }) => {
-  const { t } = useTranslation();
-  const [openDrawer, setOpenDrawer] = useState<boolean>(false);
-  const bundleTypeGlobal: boolean = bundleDetail?.type === TypeEnum.GLOBAL;
-
-  const columns: Array<Array<Array<string>>> = bundleTypeGlobal
-    ? Object.values(bundleConfigurationFields)
-    : new Array(bundleConfigurationFields.col1);
-
-  const mapColumn = (col: Array<Array<string>>, isDrawer: boolean, isFirstColumn?: boolean) =>
-    col.map((entry: Array<string>, index: number) => (
-      <Box mt={1} key={`config-detail-${entry[0]}`}>
-        {isDrawer && (!isFirstColumn || index !== 0) && <Divider />}
-        <Typography variant="body1" color="text.disabled">
-          {t(entry[1])}
-        </Typography>
-        <Typography variant="body1">
-          {formatConfigValues(bundleDetail?.[entry[0] as keyof Bundle], t)}
-        </Typography>
-      </Box>
-    ));
-
-  return (
-    <Paper elevation={3} sx={{ borderRadius: 2, padding: 3, minHeight: '310px' }}>
-      <Typography variant="overline">
-        {t('commissionBundlesPage.commissionBundleDetail.configuration')}
-      </Typography>
-      <Grid container spacing={1}>
-        {columns.map((col, i) => (
-          <Grid item xs={12 / columns.length} key={`config-grid-item-${i}`}>
-            {mapColumn(col, false)}
-          </Grid>
-        ))}
-      </Grid>
-
-      {!bundleTypeGlobal && (
-        <>
-          <ButtonNaked
-            size="large"
-            component="button"
-            onClick={() => setOpenDrawer(true)}
-            sx={{ color: 'primary.main', mt: 3 }}
-            weight="default"
-            data-testid="show-more-bundle-configuration-test"
-          >
-            + {t('general.showMore')}
-          </ButtonNaked>
-          <PaddedDrawer openDrawer={openDrawer} setOpenDrawer={setOpenDrawer}>
-            <>
-              <TitleBox
-                title={t('commissionBundlesPage.commissionBundleDetail.configuration')}
-                variantTitle="h4"
-              />
-              {Object.values(bundleConfigurationFields).map((col, i) =>
-                mapColumn(col, true, i === 0)
-              )}
-            </>
-          </PaddedDrawer>
-        </>
-      )}
-    </Paper>
-  );
-};
-
-const BundleTaxonomiesDetails = ({ bundleDetail }: { bundleDetail: Bundle }) => {
-  const { t } = useTranslation();
-  const [openDrawer, setOpenDrawer] = useState<boolean>(false);
-
-  return (
-    <Paper
-      elevation={3}
-      sx={{
-        borderRadius: 2,
-        padding: 3,
-        minHeight: '310px',
-        display: 'flex',
-        flexDirection: 'column',
-      }}
-    >
-      <Typography variant="overline">
-        {t('commissionBundlesPage.commissionBundleDetail.taxonomies')}
-      </Typography>
-
-      {bundleDetail?.transferCategoryList?.map((el, i) =>
-        i < 4 ? (
-          <Box key={`taxonomy-${el}`} mt={1}>
-            {/* TODO RETRIEVE TAXONOMIES */}
-            <Typography variant="body1" color="text.disabled">
-              {el}
-            </Typography>
-            <Typography variant="body1">{el}</Typography>
-          </Box>
-        ) : null
-      )}
-      {bundleDetail?.transferCategoryList && bundleDetail?.transferCategoryList?.length > 0 ? (
-        <>
-          <ButtonNaked
-            size="large"
-            component="button"
-            onClick={() => setOpenDrawer(true)}
-            sx={{ color: 'primary.main', mt: 'auto', justifyContent: 'start' }}
-            weight="default"
-            data-testid="show-more-bundle-taxonomies-test"
-          >
-            + {t('general.showMore')}
-          </ButtonNaked>
-          <PaddedDrawer openDrawer={openDrawer} setOpenDrawer={setOpenDrawer}>
-            <TitleBox
-              title={t('commissionBundlesPage.commissionBundleDetail.taxonomies')}
-              variantTitle="h5"
-            />
-            {bundleDetail.transferCategoryList?.map((el) => (
-              <Box key={`taxonomies-list-${el}`} mt={1}>
-                {/* TODO RETRIEVE TAXONOMIES */}
-                <Typography variant="body1" color="text.disabled">
-                  {el}
-                </Typography>
-                <Typography variant="body1">{el}</Typography>
-              </Box>
-            ))}
-          </PaddedDrawer>
-        </>
-      ) : (
-        <Alert severity="info" variant="outlined" data-testid="alert-test" sx={{ mt: 2 }}>
-          {t('commissionBundlesPage.commissionBundleDetail.noTaxonomiesAlert')}
-        </Alert>
-      )}
-    </Paper>
-  );
-};
 
 const CommissionBundleDetailPage = () => {
   const { t } = useTranslation();
@@ -316,11 +116,11 @@ const CommissionBundleDetailPage = () => {
             </Typography>
           </Grid>
           <Grid item xs={commissionBundleDetail?.type === TypeEnum.GLOBAL ? 12 : 6}>
-            <BundleConfigurationDetails bundleDetail={commissionBundleDetail} />
+            <CommissionBundleDetailConfiguration bundleDetail={commissionBundleDetail} />
           </Grid>
           {commissionBundleDetail?.type !== TypeEnum.GLOBAL && (
             <Grid item xs={6}>
-              <BundleTaxonomiesDetails bundleDetail={commissionBundleDetail} />
+              <CommissionBundleDetailTaxonomies bundleDetail={commissionBundleDetail} />
             </Grid>
           )}
         </Grid>
